scripts/requirements.txt
adal==1.2.7
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==2.0.0
anyio==3.3.4
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
avro==1.10.2
azure-core==1.17.0
azure-identity==1.6.0
azure-storage-blob==12.8.1
blessings==1.7
boto3==1.18.18
botocore==1.21.18
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cloudevents==1.2.0
colorama==0.4.4
coloredlogs==15.0.1
configparser==5.0.2
cryptography==3.4.7
deprecation==2.1.0
fastapi==0.70.0
filelock==3.3.1
google-api-core==1.29.0
google-auth==1.34.0
google-cloud-core==2.1.0
google-cloud-storage==1.41.1
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
gpustat==0.6.0
grpcio==1.38.1
h11==0.12.0
humanfriendly==10.0
idna==3.2
isodate==0.6.0
jmespath==0.10.0
joblib==0.17.0
jsonschema==4.1.0
kserve==0.7.0
kubernetes==18.20.0
minio==6.0.2
msal==1.15.0
msal-extensions==0.3.0
msgpack==1.0.2
msrest==0.6.21
multidict==5.2.0
numpy==1.18.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
opencensus==0.8.0
opencensus-context==0.1.2
packaging==21.0
portalocker==1.7.1
prometheus-client==0.11.0
protobuf==3.18.1
psutil==5.8.0
py-spy==0.3.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
ray==1.5.0
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
scikit-learn==0.20.3
scipy==1.7.1
sh==1.14.2
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
table-logger==0.3.6
tornado==6.1
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
websocket-client==1.2.1
yarl==1.7.0
